Add admin controller
