+++ Controllers/AdminController.cs
@@ -1,0 +1,48 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Profound.Data;
+
+namespace Profound.Controllers
+{
+    [Route("api/[controller]")]
+    [ApiController]
+    [Authorize(Roles = "admin")]
+    public class AdminController : ControllerBase
+    {
+        private readonly IDataRepository _dataRepository;
+
+        public AdminController(IDataRepository dataRepository)
+        {
+            _dataRepository = dataRepository;
+        }
+
+        [HttpPost("{courseId}/publish")]
+        public ActionResult PublishCourse(int courseId)
+        {
+            var course = _dataRepository.GetBaseCourse(courseId);
+            if (course != null)
+            {
+                _dataRepository.PublishCourse(courseId);
+                return Ok();
+            }
+            return NotFound();
+        }
+
+        [HttpPost("{courseId}/reject")]
+        public ActionResult RejectCourse(int courseId)
+        {
+            var course = _dataRepository.GetBaseCourse(courseId);
+            if (course != null)
+            {
+                _dataRepository.RejectCourse(courseId);
+                return Ok();
+            }
+            return NotFound();
+        }
+    }
+}
